// src/app/core/models/premiumCalculationSettingsDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */
import type { PremiumCalculationSettingsDtoPolicyCoverTable } from './premiumCalculationSettingsDtoPolicyCoverTable';
import type { PremiumCalculationSettingsDtoExtendedFamilyTable } from './premiumCalculationSettingsDtoExtendedFamilyTable';

export interface PremiumCalculationSettingsDto {
  policyCoverTable: PremiumCalculationSettingsDtoPolicyCoverTable;
  extendedFamilyTable: PremiumCalculationSettingsDtoExtendedFamilyTable;
  maxExtendedFamilyMembers: number;
}