// src/app/core/models/policyCoverRowDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */
import type { PolicyCoverRowDtoDependentCountTiersItem } from './policyCoverRowDtoDependentCountTiersItem';
import type { PolicyCoverRowDtoAgeBrackets } from './policyCoverRowDtoAgeBrackets';

export interface PolicyCoverRowDto {
  coverAmount: number;
  dependentCountTiers: PolicyCoverRowDtoDependentCountTiersItem[];
  ageBrackets: PolicyCoverRowDtoAgeBrackets;
  premium_1To5Dependents_Under65: number;
  premium_1To5Dependents_Under70: number;
  premium_1To5Dependents_Under75: number;
  premium_1To5Dependents_75Plus: number;
}