// src/app/core/models/dependentCountTierDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */
import type { DependentCountTierDtoAgeBrackets } from './dependentCountTierDtoAgeBrackets';

export interface DependentCountTierDto {
  minDependents: number;
  maxDependents: number;
  label: string;
  ageBrackets: DependentCountTierDtoAgeBrackets;
}