// src/app/core/models/extendedFamilyBenefitRowDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */

export interface ExtendedFamilyBenefitRowDto {
  minAge: number;
  maxAge: number;
  ageRange: string;
  premium_5000_Cover: number;
  premium_10000_Cover: number;
  premium_15000_Cover: number;
  premium_20000_Cover: number;
  premium_25000_Cover: number;
}